Ultrawide Module + call_aligned
